namespaced api and cors
