+++ Gemfile
@@ -13,6 +13,7 @@ gem 'jwt'
 gem 'lograge'
 gem 'mitlibraries-theme'
 gem 'puma'
+gem 'rack-cors'
 gem 'rails', '~> 5.2'
 gem 'sass-rails'
 gem 'uglifier'
