+++ app/views/api/v1/search/_extended.json.jbuilder
