+++ app/views/api/v1/search/record.json.jbuilder
