+++ config/routes.rb
@@ -3,8 +3,12 @@ Rails.application.routes.draw do
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
   root to: 'static#index'
 
-  get 'auth', to: 'auth#auth'
-  get 'search', to: 'search#search'
-  get 'record/:id', to: 'search#record', as: 'record'
-  get 'ping', to: 'search#ping'
+  namespace :api do
+    namespace :v1 do
+      get 'auth', to: 'auth#auth'
+      get 'search', to: 'search#search'
+      get 'record/:id', to: 'search#record', as: 'record'
+      get 'ping', to: 'search#ping'
+    end
+  end
 end
