+++ test/controllers/auth_controller_test.rb
@@ -4,13 +4,13 @@ class AuthControllerTest < ActionDispatch::IntegrationTest
   test 'valid credentials returns JWT' do
     u = users(:yo)
     b = Base64.encode64(u.email + ':123greetings')
-    get '/auth', headers: { 'Authorization': "Basic #{b}" }
+    get '/api/v1/auth', headers: { 'Authorization': "Basic #{b}" }
     assert_equal(200, response.status)
-    assert_equal(u.id, JWTWrapper.decode(response.body)['user_id'])
+    assert_equal(u.id, JWTWrapper.decode(JSON.parse(response.body))['user_id'])
   end
 
   test 'invalid credentials' do
-    get '/auth'
+    get '/api/v1/auth'
     assert_equal(401, response.status)
     assert_equal('{"error" : "invalid credentials"}', response.body)
   end
